requirements_jungle.txt
absl-py==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alabaster==0.7.13
astroid==2.14.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
Babel==2.13.0
blis==0.7.9
Bottleneck==1.3.5
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.7
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.6
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.5
cryptography==41.0.3
cycler==0.11.0
cymem==2.0.6
decorator==4.4.2
dill==0.3.7
docopt==0.6.2
docutils==0.18.1
exceptiongroup==1.0.4
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.0
filelock==3.9.0
flatbuffers==1.12
fonttools==4.25.0
frozenlist==1.8.0
fsspec==2023.4.0
gast==0.4.0
gmpy2==2.1.2
google-auth==2.23.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.56.2
gym==0.22.0
gym-notices==0.0.8
h5py==3.10.0
idna==3.4
imageio==2.31.5
imageio-ffmpeg==0.4.9
imagesize==1.4.1
importlib-metadata==6.8.0
importlib-resources==5.2.0
iniconfig==1.1.1
isort==5.9.3
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Preprocessing==1.1.2
Kivy==2.1.0
Kivy-Garden==0.1.5
kiwisolver==1.4.4
langcodes==3.3.0
lazy-object-proxy==1.6.0
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.9.0
Markdown==3.4.4
MarkupSafe==2.1.1
matplotlib==3.7.2
mccabe==0.7.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.7.0
munkres==1.1.4
murmurhash==1.0.7
networkx==3.1
numexpr==2.8.4
numpy==1.22.3
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.3
pathy==0.10.1
-e git+https://github.com/MaximIgenbergs/perturbation-drive@e3f8ccb51b924e6461987b8893fec65e33f555c6#egg=perturbationdrive&subdirectory=../../../external/perturbation-drive
Pillow==9.4.0
platformdirs==3.10.0
plotly==5.9.0
pluggy==1.0.0
preshed==3.0.6
prettytable==3.5.0
proglog==0.1.10
progress==1.5
propcache==0.4.1
protobuf==3.19.6
psutil==5.9.0
py==1.11.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.8
pyfiglet==0.8.post1
pygame==2.5.2
Pygments==2.16.1
pylint==2.16.2
pynmea2==1.19.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperf==2.6.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==3.13.2
python-socketio==4.5.1
pytorch-lightning==1.9.5
pytz==2022.7
PyYAML==6.0
recommonmark==0.7.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.0
shapely==2.0.7
shellingham==1.5.0
simple-pid==2.0.0
six==1.16.0
smart-open==5.2.1
snowballstemmer==2.2.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
srsly==2.4.6
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.23.0
tensorflow-estimator==2.9.0
tensorflow-macos==2.9.2
tensorflow-metal==0.5.1
termcolor==2.3.0
-e git+https://github.com/MaximIgenbergs/thesis-robust-e2e-vision-ads.git@d8dbe970b7c1edc34f7f3609e70cb6a3d534b7de#egg=thesis_robust_e2e_vision_ads
thinc==8.1.10
threadpoolctl==2.2.0
tifffile==2023.9.26
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==2.0.1
torchaudio==2.0.2
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typeguard==2.13.3
typer==0.4.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
utm==0.7.0
wasabi==0.9.1
wcwidth==0.2.5
Werkzeug==2.3.7
wrapt==1.14.1
yarl==1.22.0
zipp==3.11.0
